DPO-2816 fixes (#80)

+++ src/main/java/fi/livi/digitraffic/meri/service/portnet/location/LocationParser.java
@@ -77,7 +77,7 @@ public final class LocationParser {
         }
 
         // no sign, when 0.0
-        if(Math.abs(dValue.doubleValue()) - 0 < 0.0001) {
+        if(Math.abs(dValue) - 0 < 0.0001) {
             return 0.0;
         }
